Booking system. First edit. Implementation priority queue to request documents.

+++ src/Patron.java
@@ -14,9 +14,15 @@ import java.util.NoSuchElementException;
 public class Patron extends User {
 	/**
 	 * Patron type.
-	 * Possible values: {@code "faculty"}, {@code "student"}
+	 * Possible values: {@code "faculty"}, {@code "student"}, {@code "ta"}, {@code "vp"} , {@code "professor"}
 	 */
 	private String status;
+
+	/**
+	 * Priority of Patron.
+	 * Levels of priority: student, faculty(instructions), TA, VP, professor
+	 */
+	private int priority;
 
 	/**
 	 * List of patrons' documents IDs.
@@ -37,6 +43,7 @@ public class Patron extends User {
 	public Patron(String login, String password, String status, String name, String surname, String phone, String address) {
 		super(login, password, name, surname, phone, address);
 		this.status = status;
+		this.setPriority(status);
 	}
 
 	/**
@@ -66,6 +73,23 @@ public class Patron extends User {
 		this.status = status;
 	}
 
+	public void setPriority(String status){
+		if (status == "student"){
+			priority = 0;
+		} else if (status == "faculty"){
+			priority = 1;
+		} else if (status == "ta"){
+			priority = 2;
+		} else if (status == "vp"){
+			priority = 3;
+		} else {
+			priority = 4;
+		}
+	}
+
+	public int getPriority(){
+		return priority;
+	}
 	/**
 	 * Checks whether patron can take the following book.
 	 *
@@ -73,6 +97,8 @@ public class Patron extends User {
 	 * @param database Database that stores the information.
 	 * @return {@code true} if this patron can get the book, otherwise {@code false}.
 	 */
+
+	//TODO: NEW STATUS OF PATRON!!!
 	public boolean canRequestBook(int idBook, Database database) {
 		try {
 			database.getPatron(getId());
@@ -82,10 +108,13 @@ public class Patron extends User {
 		}
 		try {
 			Book book = database.getBook(idBook);
-			if ((this.status.toLowerCase().equals("faculty")) && (book.getNumberOfCopies() != 0) &&
+			if ((this.status.toLowerCase().equals("faculty") || (this.status.toLowerCase().equals("ta")) || (this.status.toLowerCase().equals("vp")) ||
+					(this.status.toLowerCase().equals("professor")))
+					&& (book.getNumberOfCopies() != 0) &&
 					!book.isReference() && !getListOfDocumentsPatron().contains(idBook)) {
 				return true;
-			} else if ((this.status.toLowerCase().equals("faculty"))) {
+			} else if ((this.status.toLowerCase().equals("faculty")) || (this.status.toLowerCase().equals("ta")) ||(this.status.toLowerCase().equals("vp")) ||
+					(this.status.toLowerCase().equals("professor"))) {
 				if (book.getNumberOfCopies() == 0)
 					System.out.println("Not copies");
 				if (getListOfDocumentsPatron().contains(idBook))
@@ -128,10 +157,12 @@ public class Patron extends User {
 		}
 		try {
 			JournalArticle article = database.getArticle(idArticle);
-			if ((this.status.toLowerCase().equals("faculty")) && (article.getNumberOfCopies() != 0) &&
+			if ((this.status.toLowerCase().equals("faculty")|| (this.status.toLowerCase().equals("ta")) || (this.status.toLowerCase().equals("vp"))||
+					(this.status.toLowerCase().equals("professor"))) && (article.getNumberOfCopies() != 0) &&
 					!article.isReference() && !getListOfDocumentsPatron().contains(idArticle)) {
 				return true;
-			} else if ((this.status.toLowerCase().equals("faculty"))) {
+			} else if ((this.status.toLowerCase().equals("faculty")) || (this.status.toLowerCase().equals("ta")) || (this.status.toLowerCase().equals("vp")) ||
+					(this.status.toLowerCase().equals("professor"))) {
 				if (article.getNumberOfCopies() == 0)
 					System.out.println("Not copies");
 				if (getListOfDocumentsPatron().contains(idArticle))
@@ -175,10 +206,12 @@ public class Patron extends User {
 		}
 		try {
 			AudioVideoMaterial av = database.getAV(idAV);
-			if ((this.status.toLowerCase().equals("faculty")) && (av.getNumberOfCopies() != 0) &&
+			if ((this.status.toLowerCase().equals("faculty") || (this.status.toLowerCase().equals("ta")) || (this.status.toLowerCase().equals("vp"))||
+					(this.status.toLowerCase().equals("professor"))) && (av.getNumberOfCopies() != 0) &&
 					!av.isReference() && !getListOfDocumentsPatron().contains(idAV)) {
 				return true;
-			} else if ((this.status.toLowerCase().equals("faculty"))) {
+			} else if ((this.status.toLowerCase().equals("faculty")) || (this.status.toLowerCase().equals("ta")) || (this.status.toLowerCase().equals("vp"))||
+					(this.status.toLowerCase().equals("professor"))) {
 				if (av.getNumberOfCopies() == 0)
 					System.out.println("Not copies");
 				if (getListOfDocumentsPatron().contains(idAV))
@@ -224,7 +257,8 @@ public class Patron extends User {
 		}
 		try {
 			Document doc = database.getDocument(idDocument);
-			if ((this.status.toLowerCase().equals("faculty")) && (doc.getNumberOfCopies() != 0) &&
+			if ((this.status.toLowerCase().equals("faculty") || (this.status.toLowerCase().equals("ta")) || (this.status.toLowerCase().equals("vp")) ||
+					(this.status.toLowerCase().equals("professor"))) && (doc.getNumberOfCopies() != 0) &&
 					!(doc.isReference()) && !getListOfDocumentsPatron().contains(idDocument)) {
 				return true;
 			} else if (doc.isAllowedForStudents() &&
@@ -249,33 +283,33 @@ public class Patron extends User {
 		}
 	}
 
-
-
-
 	/**
 	 * Take the following book ang give it to this patron.
 	 *
 	 * @param idBook   Book to take.
 	 * @param database Database that stores the information.
 	 */
+
 	public void takeBook(int idBook, Database database) {
 		try {
-			getListOfDocumentsPatron().add(idBook);
-			database.getBook(idBook).deleteCopy();
-			decreaseCountOfCopies(idBook, database);
-			Date dateBook = new Date();
-			Date dateExpire = new Date();
-			if (database.getBook(idBook).isBestseller())
-				dateExpire.setTime(dateExpire.getTime() + 14 * 24 * 60 * 60 * 1000);
-			else {
-				if (getStatus().toLowerCase().equals("student"))
-					dateExpire.setTime(dateExpire.getTime() + 21 * 24 * 60 * 60 * 1000);
-				else if (getStatus().toLowerCase().equals("faculty")){
-					dateExpire.setTime(dateExpire.getTime() + 28L * 24 * 60 * 60 * 1000);
-				} else {
-					dateExpire.setTime(dateExpire.getTime() + 7 * 24 * 60 * 60 * 1000);
+			if (canRequestBook(idBook, database)) {
+				this.getListOfDocumentsPatron().add(idBook);
+				database.getBook(idBook).deleteCopy();
+				decreaseCountOfCopies(idBook, database);
+				Date dateBook = new Date();
+				Date dateExpire = new Date();
+				if (database.getBook(idBook).isBestseller())
+					dateExpire.setTime(dateExpire.getTime() + 14 * 24 * 60 * 60 * 1000);
+				else {
+					if (status.toLowerCase().equals("student"))
+						dateExpire.setTime(dateExpire.getTime() + 21 * 24 * 60 * 60 * 1000);
+					else if (status.toLowerCase().equals("faculty") ||(status.toLowerCase().equals("ta")) ||
+							(status.toLowerCase().equals("professor"))){
+						dateExpire.setTime(dateExpire.getTime() + 28L * 24 * 60 * 60 * 1000);
+					} else {
+						dateExpire.setTime(dateExpire.getTime() + 7 * 24 * 60 * 60 * 1000);
+					}
 				}
-
 
 				Debt debt = new Debt(getId(), idBook, dateBook, dateExpire, 0, true);
 				database.insertDebt(debt);
@@ -293,16 +327,18 @@ public class Patron extends User {
 	 */
 	public void takeAV(int idAV, Database database) {
 		try {
-			getListOfDocumentsPatron().add(idAV);
-			database.getAV(idAV).deleteCopy();
-			decreaseCountOfCopies(idAV, database);
-			Date date = new Date();
-			Date date2 = new Date();
-			if(getStatus().toLowerCase().equals("vp"))
-				date2.setTime(date2.getTime() + 7 * 24 * 60 * 60 * 1000);
-			else date2.setTime(date2.getTime() + 14 * 24 * 60 * 60 * 1000);
-			Debt debt = new Debt(getId(), idAV, date, date2, 0, true);
-			database.insertDebt(debt);
+			if (canRequestAV(idAV, database)) {
+				this.getListOfDocumentsPatron().add(idAV);
+				database.getAV(idAV).deleteCopy();
+				decreaseCountOfCopies(idAV, database);
+				Date date = new Date();
+				Date date2 = new Date();
+				if(status.toLowerCase().equals("vp"))
+					date2.setTime(date2.getTime() + 7 * 24 * 60 * 60 * 1000);
+					else date2.setTime(date2.getTime() + 14 * 24 * 60 * 60 * 1000);
+				Debt debt = new Debt(getId(), idAV, date, date2, 0, true);
+				database.insertDebt(debt);
+			}
 		} catch (SQLException | NoSuchElementException e) {
 			System.out.println("Incorrect id" + idAV);
 		}
@@ -316,21 +352,21 @@ public class Patron extends User {
 	 */
 	public void takeArticle(int idArticle, Database database) throws ParseException {
 		try {
-			getListOfDocumentsPatron().add(idArticle);
-			database.getArticle(idArticle).deleteCopy();
-			decreaseCountOfCopies(idArticle, database);
-			Date date = new Date();
-			Date date2 = new Date();
-			if(getStatus().toLowerCase().equals("vp"))
-				date2.setTime(date2.getTime() + 7 * 24 * 60 * 60 * 1000);
-			else date2.setTime(date2.getTime() + 14 * 60 * 60 * 1000 * 24);
-			Debt debt = new Debt(getId(), idArticle, date, date, 0, true);
+			if (canRequestArticle(idArticle, database)) {
+				this.getListOfDocumentsPatron().add(idArticle);
+				database.getArticle(idArticle).deleteCopy();
+				decreaseCountOfCopies(idArticle, database);
+				Date date = new Date();
+				Date date2 = new Date();
+				if(status.toLowerCase().equals("vp"))
+					date2.setTime(date2.getTime() + 7 * 24 * 60 * 60 * 1000);
+				else date2.setTime(date2.getTime() + 14 * 60 * 60 * 1000 * 24);
+				Debt debt = new Debt(getId(), idArticle, date, date, 0, true);
 
-			database.insertDebt(debt);
+				database.insertDebt(debt);
+			}
 		} catch (SQLException | NoSuchElementException e) {
 			System.out.println("Incorrect id" + idArticle);
-		} catch(ParseException e){
-
 		}
 	}
 
@@ -341,64 +377,59 @@ public class Patron extends User {
 	 * @param database   Database that stores the information.
 	 */
 	public void takeDocument(int idDocument, Database database) {
-		try{
-			Document doc= database.getDocument(idDocument);
-			getListOfDocumentsPatron().add(idDocument);
-			doc.deleteCopy();
-			decreaseCountOfCopies(idDocument, database);
-			Date date = new Date();
-			Debt debt = new Debt(getId(), idDocument, date, date, 0, true);
-			database.insertDebt(debt);
-		} catch(SQLException | NoSuchElementException e){
-			System.out.println("Incorrect ID!");
-		}
-	}
-
-
-
-	public void makeReturnRequest(int debtID, Database database){
-		try{
-			Debt debt = database.getDebt(debtID);
-			debt.countFee(database);
-			if(debt.getFee() > 0)
-				System.out.println("You cannot return this document until you pay for delay!");
-			else{
-				ReturnRequest request = new ReturnRequest(debtID, database);
+		try {
+			if (canRequestDocument(idDocument, database)) {
+				getListOfDocumentsPatron().add(idDocument);
+				database.getDocument(idDocument).deleteCopy();
+				decreaseCountOfCopies(idDocument, database);
+				Date date = new Date();
+				Debt debt = new Debt(getId(), idDocument, date, date, 0, true);
+				database.insertDebt(debt);
 			}
-		} catch (SQLException e){
-
-		} catch (ParseException e){
-
+		} catch (SQLException | NoSuchElementException e) {
+			System.out.println("Incorrect id" + idDocument);
 		}
-
 	}
-
 
 	/**
-	 * Return the document to the library.
+	 * Make request the following document
+	 * @param idDocument document to request
+	 * @param database DatabasE that stores the information
+	 * @throws SQLException
+	 */
+	public void makeRequest(int idDocument, Database database) throws SQLException {
+		try {
+			//TODO: set current date
+			Date currentDate = new Date();
+			Request request = new Request(this.getId(), idDocument, currentDate, database);
+			request.addToQueue(this.getId(), idDocument, database);
+		} catch(NoSuchElementException e ){
+			System.out.println("Incorrect id" + idDocument);
+		}
+	}
+
+	/**
+	 * Decrease the number of copies of specified document by one.
 	 *
 	 * @param idDocument Document ID.
 	 * @param database   Database that stores the information.
+	 * @throws SQLException If passed the wrong document ID.
 	 */
-	public void returnDocument(int idDocument, Database database) {
-		try {
+	private void decreaseCountOfCopies(int idDocument, Database database) throws SQLException {
+		int count = database.getDocument(idDocument).getNumberOfCopies();
+		database.editDocumentColumn(idDocument, "num_of_copies", Integer.toString(count - 1));
+	}
 
-
-			for (int i = 0; i < getListOfDocumentsPatron().size(); i++) {
-				if (getListOfDocumentsPatron().get(i).equals(idDocument)) {
-					getListOfDocumentsPatron().remove(i);
-					break;
-				}
-			}
-			database.getDocument(idDocument).addCopy();
-			increaseCountOfCopies(idDocument, database);
-			int debtID = database.findDebtID(this.getId(), idDocument);
-			database.deleteDebt(debtID);
-		} catch (NoSuchElementException | SQLException e) {
-			System.out.println("Incorrect id");
-		} catch (IndexOutOfBoundsException e) {
-			System.out.println("Incorrect input");
-		}
+	/**
+	 * Increase the number of copies of specified document by one.
+	 *
+	 * @param idDocument Document ID.
+	 * @param database   Database that stores the information.
+	 * @throws SQLException If passed the wrong document ID.
+	 */
+	private void increaseCountOfCopies(int idDocument, Database database) throws SQLException {
+		int count = database.getDocument(idDocument).getNumberOfCopies();
+		database.editDocumentColumn(idDocument, "num_of_copies", Integer.toString(count + 1));
 	}
 
 	/**
@@ -409,16 +440,15 @@ public class Patron extends User {
 	 */
 	public void returnBook(int idBook, Database database) {
 		try {
-			Book book = database.getBook(idBook);
-			for (int i = 0; i < getListOfDocumentsPatron().size(); i++) {
+			for (int i = 0; i < listOfDocumentsPatron.size(); i++) {
 				if (getListOfDocumentsPatron().get(i).equals(idBook)) {
 					getListOfDocumentsPatron().remove(i);
 					break;
 				}
 			}
-			book.addCopy();
+			database.getBook(idBook).addCopy();
 			increaseCountOfCopies(idBook, database);
-			int debtID = database.findDebtID(getId(), idBook);
+			int debtID = database.findDebtID(this.getId(), idBook);
 			database.deleteDebt(debtID);
 		} catch (NoSuchElementException | SQLException e) {
 			System.out.println("Incorrect id");
@@ -435,23 +465,20 @@ public class Patron extends User {
 	 */
 	public void returnArticle(int idArticle, Database database) throws ParseException {
 		try {
-			JournalArticle article = database.getArticle(idArticle);
-			for (int i = 0; i < getListOfDocumentsPatron().size(); i++) {
+			for (int i = 0; i < listOfDocumentsPatron.size(); i++) {
 				if (getListOfDocumentsPatron().get(i).equals(idArticle)) {
 					getListOfDocumentsPatron().remove(i);
 					break;
 				}
 			}
-			article.addCopy();
+			database.getArticle(idArticle).addCopy();
 			increaseCountOfCopies(idArticle, database);
-			int debtID = database.findDebtID(getId(), idArticle);
+			int debtID = database.findDebtID(this.getId(), idArticle);
 			database.deleteDebt(debtID);
 		} catch (NoSuchElementException | SQLException e) {
 			System.out.println("Incorrect id");
 		} catch (IndexOutOfBoundsException e) {
 			System.out.println("Incorrect input");
-		} catch (ParseException e){
-
 		}
 	}
 
@@ -463,7 +490,7 @@ public class Patron extends User {
 	 */
 	public void returnAV(int idAV, Database database) {
 		try {
-			for (int i = 0; i < getListOfDocumentsPatron().size(); i++) {
+			for (int i = 0; i < listOfDocumentsPatron.size(); i++) {
 				if (getListOfDocumentsPatron().get(i).equals(idAV)) {
 					getListOfDocumentsPatron().remove(i);
 					break;
@@ -471,7 +498,7 @@ public class Patron extends User {
 			}
 			database.getAV(idAV).addCopy();
 			increaseCountOfCopies(idAV, database);
-			int debtID = database.findDebtID(getId(), idAV);
+			int debtID = database.findDebtID(this.getId(), idAV);
 			database.deleteDebt(debtID);
 		} catch (NoSuchElementException | SQLException e) {
 			System.out.println("Incorrect id");
@@ -480,10 +507,30 @@ public class Patron extends User {
 		}
 	}
 
-
-
-
-
+	/**
+	 * Return the document to the library.
+	 *
+	 * @param idDocument Document ID.
+	 * @param database   Database that stores the information.
+	 */
+	public void returnDocument(int idDocument, Database database) {
+		try {
+			for (int i = 0; i < listOfDocumentsPatron.size(); i++) {
+				if (getListOfDocumentsPatron().get(i).equals(idDocument)) {
+					getListOfDocumentsPatron().remove(i);
+					break;
+				}
+			}
+			database.getDocument(idDocument).addCopy();
+			increaseCountOfCopies(idDocument, database);
+			int debtID = database.findDebtID(this.getId(), idDocument);
+			database.deleteDebt(debtID);
+		} catch (NoSuchElementException | SQLException e) {
+			System.out.println("Incorrect id");
+		} catch (IndexOutOfBoundsException e) {
+			System.out.println("Incorrect input");
+		}
+	}
 
 	/**
 	 * Compare this patron with another.
@@ -536,29 +583,4 @@ public class Patron extends User {
 			System.out.println("By default");
 		}
 	}
-
-	/**
-	 * Decrease the number of copies of specified document by one.
-	 *
-	 * @param idDocument Document ID.
-	 * @param database   Database that stores the information.
-	 * @throws SQLException If passed the wrong document ID.
-	 */
-	private void decreaseCountOfCopies(int idDocument, Database database) throws SQLException {
-		int count = database.getDocument(idDocument).getNumberOfCopies();
-		database.editDocumentColumn(idDocument, "num_of_copies", Integer.toString(count - 1));
-	}
-
-	/**
-	 * Increase the number of copies of specified document by one.
-	 *
-	 * @param idDocument Document ID.
-	 * @param database   Database that stores the information.
-	 * @throws SQLException If passed the wrong document ID.
-	 */
-	private void increaseCountOfCopies(int idDocument, Database database) throws SQLException {
-		int count = database.getDocument(idDocument).getNumberOfCopies();
-		database.editDocumentColumn(idDocument, "num_of_copies", Integer.toString(count + 1));
-	}
-
 }
